refactor: CartPage내에 calculateCartTotal 함수를 models/cart 내부로 분리
