+++ src/origin/components/AdminPage.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 
-import { Coupon, Discount, Product } from '../../types';
+import type { Coupon, Discount, Product } from '../../types';
 
 interface Props {
   products: Product[];
